minor server side refactor
Mostly moving around things across modules. No change in
functionality.
